Benchmark some serialization functions with the newly added `com.huanshankeji.benchmark.kotlinx-benchmark-multiplatform-conventions` plugin
The results show that serializing with a static serializer does improve performance. The performance is brought up to 2.5 times in the case of JSON and 5 times in the case of Protobuf. Surprisingly in the case without a static serializer, the performance of serializing to Protobuf is 5% less than that of JSON, which is might be caused by more complex generation of Protobuf serializers and that the data contains a string only instead of any numbers. The nothing workaround serialization function is indeed very slow, with its performance being only 1/4 of the origin function.

The JS benchmark doesn't run yet, which appears to result from a bug.

A corresponding commit: https://github.com/huanshankeji/gradle-common/commit/57e43efd5aff6daf9629ca58fea2f17b154bdba2